requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.8.0
oauthlib==3.2.2
openai==1.55.2
outcome==1.3.0.post0
pydantic==2.10.2
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
schedule==1.2.2
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
